requirements.txt
# Core
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3

# Data handling
pandas==2.2.2

# Supabase
supabase==2.4.0

# OpenAI
openai==1.42.0

# Dashboard
streamlit==1.37.1

# Deployment
modal==0.63.53
